backend/users/urls_auth.py
from django.urls import path
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

from .views import ChangePasswordView, MeView, RegisterView, UpdateDisplayNameView

urlpatterns = [
    path("register/", RegisterView.as_view(), name="register"),
    path("login/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
    path("me/", MeView.as_view(), name="me"),
    path("me/display-name/", UpdateDisplayNameView.as_view(), name="update_display_name"),
    path("me/password/", ChangePasswordView.as_view(), name="change_password"),
]
