backend/requirements.txt
django>=5.2.6
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.4.0
psycopg2-binary>=2.9.9
django-environ>=0.11.2
